[Refactor/#29] chat 패키지 위치 변경
