+++ src/main/java/com/back/domain/chat/chat/entity/Chat.java
@@ -1,4 +1,4 @@
-package com.back.domain.chat.entity;
+package com.back.domain.chat.chat.entity;
 
 import com.back.global.jpa.entity.BaseEntity;
 import jakarta.persistence.*;
