+++ src/main/java/com/back/domain/chat/chat/entity/ChatRoomType.java
@@ -1,4 +1,4 @@
-package com.back.domain.chat.entity;
+package com.back.domain.chat.chat.entity;
 
 public enum ChatRoomType {
     AUCTION, POST
